+++ packages/ui/select/index.js
@@ -1,2 +1,4 @@
-export const Select = defineAsyncComponent(() => import('./select.vue'))
-export const SelectOption = defineAsyncComponent(() => import('./select-option.vue'))
+import { defineAsyncComponent } from "vue";
+
+export const Select = defineAsyncComponent(() => import("./select.vue"));
+export const SelectOption = defineAsyncComponent(() => import("./select-option.vue"));