+++ packages/codepen/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @lil-el/codepen
 
+## 1.2.1
+
+### Patch Changes
+
+- set default theme color
+- Updated dependencies
+  - @lil-el/ui@1.1.1
+
 ## 1.2.0
 
 ### Minor Changes