+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lil-el/ui
 
+## 1.1.1
+
+### Patch Changes
+
+- set default theme color
+
 ## 1.1.0
 
 ### Minor Changes